projects/incident_investigation/requirements.txt
# AI-Powered Incident Investigation System - Requirements

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.1
python-docx>=1.1.0  # For Word document generation in testsre Dependencies
openai>=1.0.0
python-dotenv>=1.0.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Document Processing
PyPDF2>=3.0.0
python-docx>=1.0.0
openpyxl>=3.1.0
Pillow>=10.0.0

# OCR and Vision
pytesseract>=0.3.10

# NLP and Embeddings
spacy>=3.7.0
sentence-transformers>=2.2.2

# Vector Database
chromadb>=0.4.0
# Alternative: pinecone-client>=2.2.0

# Diagram Generation
graphviz>=0.20.1
matplotlib>=3.7.0
seaborn>=0.12.0
networkx>=3.1

# Report Generation
reportlab>=4.0.0
Jinja2>=3.1.2

# Web Framework (for API)
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.4.0

# Database
psycopg2-binary>=2.9.9
SQLAlchemy>=2.0.0
redis>=5.0.0

# Task Queue
celery>=5.3.0
# Alternative: rq>=1.15.0

# File Storage
minio>=7.2.0
# Alternative: boto3>=1.29.0

# Utilities
python-dateutil>=2.8.0
pytz>=2023.3

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
pytest-asyncio>=0.21.0
httpx>=0.25.0

# Code Quality
black>=23.11.0
flake8>=6.1.0
mypy>=1.7.0

# Monitoring and Logging
prometheus-client>=0.19.0
sentry-sdk>=1.38.0

# Security
cryptography>=41.0.0
pyjwt>=2.8.0

# Development
ipython>=8.17.0
jupyter>=1.0.0
